+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>US States Quiz</title>
     <link rel="stylesheet" href="styles.css" />
-    <script src="./js/index.js" defer></script>
+    <script src="./js/index.js" type="module"></script>
   </head>
   <body>
     <header>
